requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.8
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
frozenlist==1.6.2
idna==3.10
lxml==5.4.0
multidict==6.4.4
propcache==0.3.1
pydantic==2.11.5
pydantic-xml==2.17.0
pydantic_core==2.33.2
python-dotenv==1.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
yarl==1.20.0
